test: refactor SafeTx def test cases into their own suite
